Age Projection 100
"AgeProjection100" is a Python project that allows users to input their name and age and then calculates their age in 100 years as well as the year they will turn 100. It provides a simple interface for users to interact with, making it easy to use for anyone interested in knowing these projections.